pyproject.toml
[project]
name = "tmux-trainsh"
version = "26.313.2342"
description = "GPU training workflow automation with tmux"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = ["cryptography>=42", "pyyaml>=6.0"]

[project.scripts]
train = "trainsh.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["trainsh"]
